Centralize evidence engine goal config
