+++ scripts/data/validate-sleep-evidence-engine.mjs
@@ -4,6 +4,7 @@ import fs from 'node:fs'
 import path from 'node:path'
 import { fileURLToPath } from 'node:url'
 import { validateEvidenceEnginePayload } from './evidence-engine-validation.mjs'
+import { getEvidenceEngineGoalConfigs } from './evidence-engine-goals.mjs'
 
 const __filename = fileURLToPath(import.meta.url)
 const __dirname = path.dirname(__filename)
@@ -13,60 +14,31 @@ const dataDir = process.argv.includes('--data-dir')
   ? process.argv[process.argv.indexOf('--data-dir') + 1]
   : (process.argv.find((arg) => arg.startsWith('--data-dir='))?.slice(11) || 'public/data')
 
-const payloadPaths = {
-  sleep: path.resolve(repoRoot, dataDir, 'evidence-engine', 'sleep.json'),
-  stress: path.resolve(repoRoot, dataDir, 'evidence-engine', 'stress.json'),
-}
-
-const SLEEP_PROBLEMS = new Set([
-  'sleep_onset',
-  'sleep_quality',
-  'night_waking',
-  'racing_mind',
-  'relaxation',
-])
-
 function main() {
-  if (!fs.existsSync(payloadPaths.sleep)) {
-    throw new Error(`[sleep-evidence-engine] missing generated payload: ${path.relative(repoRoot, payloadPaths.sleep)}`)
+  for (const config of getEvidenceEngineGoalConfigs()) {
+    const payloadPath = path.resolve(repoRoot, dataDir, 'evidence-engine', `${config.goal}.json`)
+    if (!fs.existsSync(payloadPath)) {
+      throw new Error(`[${config.goal}-evidence-engine] missing generated payload: ${path.relative(repoRoot, payloadPath)}`)
+    }
+
+    const payload = JSON.parse(fs.readFileSync(payloadPath, 'utf8'))
+    const validProblems = new Set(
+      Object.keys(payload?.problemLabels || {}).filter(Boolean)
+    )
+    const errors = validateEvidenceEnginePayload(payload, {
+      goal: config.goal,
+      problemField: config.problemField,
+      validProblems: validProblems.size > 0 ? validProblems : config.validProblems,
+    })
+
+    if (errors.length > 0) {
+      console.error(`[${config.goal}-evidence-engine] validation failed`)
+      for (const error of errors) console.error(`- ${error}`)
+      process.exit(1)
+    }
+
+    console.log(`[${config.goal}-evidence-engine] validation OK: ${payload.claims.length} claims, ${payload.safetyNotes.length} safety notes`)
   }
-
-  const sleepPayload = JSON.parse(fs.readFileSync(payloadPaths.sleep, 'utf8'))
-  const sleepErrors = validateEvidenceEnginePayload(sleepPayload, {
-    goal: 'sleep',
-    problemField: 'sleep_problem',
-    validProblems: SLEEP_PROBLEMS,
-  })
-
-  if (sleepErrors.length > 0) {
-    console.error('[sleep-evidence-engine] validation failed')
-    for (const error of sleepErrors) console.error(`- ${error}`)
-    process.exit(1)
-  }
-
-  console.log(`[sleep-evidence-engine] validation OK: ${sleepPayload.claims.length} claims, ${sleepPayload.safetyNotes.length} safety notes`)
-
-  if (!fs.existsSync(payloadPaths.stress)) {
-    throw new Error(`[stress-evidence-engine] missing generated payload: ${path.relative(repoRoot, payloadPaths.stress)}`)
-  }
-
-  const stressPayload = JSON.parse(fs.readFileSync(payloadPaths.stress, 'utf8'))
-  const stressProblems = new Set(
-    Object.keys(stressPayload?.problemLabels || {}).filter(Boolean)
-  )
-  const stressErrors = validateEvidenceEnginePayload(stressPayload, {
-    goal: 'stress',
-    problemField: 'stress_problem',
-    validProblems: stressProblems,
-  })
-
-  if (stressErrors.length > 0) {
-    console.error('[stress-evidence-engine] validation failed')
-    for (const error of stressErrors) console.error(`- ${error}`)
-    process.exit(1)
-  }
-
-  console.log(`[stress-evidence-engine] validation OK: ${stressPayload.claims.length} claims, ${stressPayload.safetyNotes.length} safety notes`)
 }
 
 main()
